requirements.txt
numpy
pandas
scikit-learn
entsoe-py
tensorflow
python-dotenv
redis
codecarbon
flask